lazy initialization of static fields
This complexifies a bit the code generation and the generated code, but
it allows the code to be robust to classes order. When bundling many
classes together, or when a file defines a lot of classes/inner classes,
the order in which they appear does not matter anymore w.r.t static
fields, which mimics much better the Java semantics.