avoid inner class name clashes
TypeScript checks that two inner classes of the same name are compatible
if they belong to classes that have a subclassing link. IMO it is
useless and it is probably a bug. But still, we have to rename the
classes to avoid these clashes and allow TypeScript to compile the
generated code.