Support Map initialization via non-Map arguments
This patch ignores any arguments that isn't of Map type during
the invocation of a Map collection. Based on the supported
data-structures, these non-Map arguments can either be of type
int (capacity, loadFactor) or Comparator. The use case for the
latter is not currently supported while the former have no equivalent
in javascript.

Fixes #508